- debugging

+++ src/main/java/org/jlab/ersap/tridas/ana/DataFile2HipoBank.java
@@ -117,14 +117,15 @@ public class DataFile2HipoBank {
 
         while (dfh.dataBuffer.position() < dfh.dataBuffer.limit()) {
             int c = dfh.dataBuffer.getInt();
-            long d = dfh.dataBuffer.getLong();
+//            long d = dfh.dataBuffer.getLong();
+            int d = dfh.dataBuffer.getInt();
 
             int channel = bits(c,0,4);
             int slot = bits(c,4,5);
             int crate = bits(c,9,7);
             int charge = bits(c,16,16);
-            long time = bits(d, 0,16);
-            long frame_count = bits(d,16,48);
+            int time = bits(d, 0,16);
+            int frame_count = bits(d,16,48);
             System.out.println("DDD " +
                     "channel = " + channel
                     + " slot = " + slot
